+++ test/interactivityServiceTest.ts
@@ -24,25 +24,38 @@
  *  THE SOFTWARE.
  */
 import powerbi from "powerbi-visuals-api";
-import { SelectableDataPoint,  InteractivityService, createInteractivityService} from "../src/interactivityService";
+import {
+    IBehaviorOptions
+} from "../src/interactivityBaseService";
+
+import {
+    createInteractivitySelectionService,
+    InteractivitySelectionService,
+    SelectionDataPoint
+} from "../src/interactivitySelectionService";
 import { MockBehavior } from "./mocks/mockInteractiveBehavior";
 import { createVisualHost, createSelectionId} from "powerbi-visuals-utils-testutils";
 // powerbi.extensibility
 import ISelectionId = powerbi.visuals.ISelectionId;
 import IVisualHost = powerbi.extensibility.visual.IVisualHost;
-import ISelectionManager = powerbi.extensibility.ISelectionManager;
+
+interface ChicletSlicerBehaviorOptions extends IBehaviorOptions<SelectionDataPoint> {
+    some: string;
+    collection: string;
+    random: string;
+}
 
 describe("Interactivity service", () => {
     let host: IVisualHost,
-        interactivityService: InteractivityService,
+        interactivityService: InteractivitySelectionService,
         identities: ISelectionId[],
-        selectableDataPoints: SelectableDataPoint[],
+        selectableDataPoints: SelectionDataPoint[],
         behavior: MockBehavior;
 
     beforeEach(() => {
         host = createVisualHost();
 
-        interactivityService = createInteractivityService(host) as InteractivityService;
+        interactivityService = createInteractivitySelectionService(host) as InteractivitySelectionService;
 
         identities = [
             createSelectionId(),
@@ -52,7 +65,7 @@ describe("Interactivity service", () => {
             createSelectionId(),
             createSelectionId()
         ];
-        selectableDataPoints = <SelectableDataPoint[]>[
+        selectableDataPoints = <SelectionDataPoint[]>[
             { selected: false, identity: identities[0] },
             { selected: false, identity: identities[1] },
             { selected: false, identity: identities[2] },
@@ -64,33 +77,12 @@ describe("Interactivity service", () => {
         behavior = new MockBehavior(selectableDataPoints);
     });
 
-    describe("applySelectionFilter", () => {
-        it("shouldn't throw any exceptions if the selectionManager is undefined", () => {
-            interactivityService["selectionManager"] = null;
-
-            expect(() => {
-                interactivityService.applySelectionFilter();
-            }).not.toThrow();
-        });
-
-        it("the selectionManager.applySelectionFilter should be called", () => {
-            (interactivityService["selectionManager"] as ISelectionManager).applySelectionFilter = () => { };
-
-            spyOn(interactivityService["selectionManager"], "applySelectionFilter");
-
-            interactivityService.applySelectionFilter();
-
-            expect((interactivityService["selectionManager"] as ISelectionManager).applySelectionFilter)
-                .toHaveBeenCalled();
-        });
-    });
-
     describe("Binding", () => {
 
         it("Basic binding", () => {
             spyOn(behavior, "bindEvents");
             spyOn(behavior, "renderSelection");
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
             expect(behavior.bindEvents).toHaveBeenCalled();
             expect(behavior.verifyCleared()).toBeTruthy();
             expect(behavior.renderSelection).not.toHaveBeenCalled();
@@ -99,12 +91,14 @@ describe("Interactivity service", () => {
 
         it("Binding passes behaviorOptions", () => {
             spyOn(behavior, "bindEvents");
-            let arbitraryBehaviorOptions = {
+            const arbitraryBehaviorOptions: ChicletSlicerBehaviorOptions = {
+                dataPoints: selectableDataPoints,
+                behavior: behavior,
                 some: "random",
                 collection: "of",
                 random: "stuff",
             };
-            interactivityService.bind(selectableDataPoints, behavior, arbitraryBehaviorOptions);
+            interactivityService.bind(arbitraryBehaviorOptions);
             expect(behavior.bindEvents).toHaveBeenCalledWith(arbitraryBehaviorOptions, interactivityService);
         });
     });
@@ -113,7 +107,7 @@ describe("Interactivity service", () => {
 
         it("Basic selection", () => {
             spyOn(behavior, "renderSelection");
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
             behavior.selectIndex(0, false);
             expect(behavior.verifySingleSelectedAt(0)).toBeTruthy();
             expect(behavior.renderSelection).toHaveBeenCalledWith(true);
@@ -121,15 +115,15 @@ describe("Interactivity service", () => {
         });
 
         it("Apply selection", () => {
-            let newDataPoints = selectableDataPoints.map((selectableDataPoint: SelectableDataPoint) => {
+            let newDataPoints = selectableDataPoints.map((selectableDataPoint: SelectionDataPoint) => {
                 return {
                     selected: false,
                     identity: selectableDataPoint.identity
-                } as SelectableDataPoint;
+                } as SelectionDataPoint;
             });
 
             spyOn(behavior, "renderSelection");
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
             behavior.selectIndex(0, false);
             expect(behavior.verifySingleSelectedAt(0)).toBeTruthy();
             expect(behavior.renderSelection).toHaveBeenCalledWith(true);
@@ -144,7 +138,7 @@ describe("Interactivity service", () => {
 
         it("Clear selection through event", () => {
             spyOn(behavior, "renderSelection");
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
             behavior.selectIndex(0, false);
             behavior.clear();
             expect(behavior.verifyCleared()).toBeTruthy();
@@ -154,7 +148,7 @@ describe("Interactivity service", () => {
 
         it("Clear selection through service", () => {
             spyOn(behavior, "renderSelection");
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
             behavior.selectIndex(0, false);
             interactivityService.clearSelection();
             expect(behavior.verifyCleared()).toBeTruthy();
@@ -163,15 +157,52 @@ describe("Interactivity service", () => {
         });
 
         it("Multiple single selects", () => {
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
             for (let i = 0, ilen = selectableDataPoints.length; i < ilen; i++) {
                 behavior.selectIndex(i, false);
                 expect(behavior.verifySingleSelectedAt(i)).toBeTruthy();
             }
         });
 
+        describe("Multiple selects", () => {
+            it("should select all of data points if multiSelect is false", () => {
+                interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
+                interactivityService.handleSelection(selectableDataPoints, false);
+
+                for (let dataPointIndex = 0; dataPointIndex < selectableDataPoints.length; dataPointIndex++) {
+                    expect(behavior.verifySingleSelectedAt(dataPointIndex)).toBeTruthy();
+                }
+            });
+
+            it("should select all of data point if multiSelect is true and dataPoints are applied by small groups", () => {
+                interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
+
+                const amountOfDataPoints: number = selectableDataPoints.length;
+
+                const firstGroup: SelectionDataPoint[] = selectableDataPoints.slice(0, amountOfDataPoints / 2);
+                const secondGroup: SelectionDataPoint[] = selectableDataPoints.slice(amountOfDataPoints / 2, amountOfDataPoints);
+
+                interactivityService.handleSelection(firstGroup, true);
+                interactivityService.handleSelection(secondGroup, true);
+
+                for (let dataPointIndex = 0; dataPointIndex < selectableDataPoints.length; dataPointIndex++) {
+                    expect(behavior.verifySingleSelectedAt(dataPointIndex)).toBeTruthy();
+                }
+            });
+        });
+
+        it("should select nothing if dataPoints are empty", () => {
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
+
+            interactivityService.handleSelection(null, null);
+
+            for (let dataPointIndex = 0; dataPointIndex < selectableDataPoints.length; dataPointIndex++) {
+                expect(behavior.verifySingleSelectedAt(dataPointIndex)).toBeFalsy();
+            }
+        });
+
         it("Single select clears", () => {
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
             behavior.selectIndex(1, false);
             expect(behavior.verifySingleSelectedAt(1)).toBeTruthy();
             behavior.selectIndex(1, false);
@@ -179,7 +210,11 @@ describe("Interactivity service", () => {
         });
 
         it("Single select null identity does not crash", () => {
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({
+                dataPoints:
+                selectableDataPoints,
+                behavior: behavior
+            });
             behavior.select({
                 identity: createSelectionId(),
                 selected: false,
@@ -188,7 +223,7 @@ describe("Interactivity service", () => {
         });
 
         it("Basic multiselect", () => {
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
             behavior.selectIndex(1, true);
             expect(behavior.verifySelectionState([false, true, false, false, false, false, false])).toBeTruthy();
             behavior.selectIndex(2, true);
@@ -198,7 +233,7 @@ describe("Interactivity service", () => {
         });
 
         it("Multiselect clears", () => {
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
             behavior.selectIndex(1, true);
             expect(behavior.verifySelectionState([false, true, false, false, false, false, false])).toBeTruthy();
             behavior.selectIndex(2, true);
@@ -212,7 +247,7 @@ describe("Interactivity service", () => {
         });
 
         it("Single and multiselect", () => {
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
             behavior.selectIndex(1, false);
             expect(behavior.verifySingleSelectedAt(1)).toBeTruthy();
             behavior.selectIndex(2, true);
@@ -226,7 +261,7 @@ describe("Interactivity service", () => {
         });
 
         it("Null identity", () => {
-            let nullIdentity: SelectableDataPoint = {
+            let nullIdentity: SelectionDataPoint = {
                 selected: false,
                 identity: null,
                 specificIdentity: createSelectionId(),
@@ -236,7 +271,7 @@ describe("Interactivity service", () => {
         });
 
         it("Null specific identity", () => {
-            let nullIdentity: SelectableDataPoint = {
+            let nullIdentity: SelectionDataPoint = {
                 selected: false,
                 identity: createSelectionId(),
                 specificIdentity: null,
@@ -246,7 +281,7 @@ describe("Interactivity service", () => {
         });
 
         it("Null for identity and specific identity", () => {
-            let nullIdentity: SelectableDataPoint = {
+            let nullIdentity: SelectionDataPoint = {
                 selected: false,
                 identity: null,
                 specificIdentity: null,
@@ -264,8 +299,8 @@ describe("Interactivity service", () => {
                 { selected: false, identity: selectableDataPoints[1].identity },
             ];
             let legendBehavior = new MockBehavior(legendDataPoints);
-            interactivityService.bind(selectableDataPoints, behavior, null);
-            interactivityService.bind(legendDataPoints, legendBehavior, null, { isLegend: true });
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
+            interactivityService.bind({ dataPoints: legendDataPoints, behavior: legendBehavior, interactivityServiceOptions: { isLegend: true } });
 
             legendBehavior.selectIndex(0);
             expect(legendBehavior.verifySingleSelectedAt(0)).toBeTruthy();
@@ -286,7 +321,7 @@ describe("Interactivity service", () => {
                 { selected: false, identity: createSelectionId() },
             ];
             behavior = new MockBehavior(selectableDataPoints);
-            interactivityService.bind(selectableDataPoints, behavior, null);
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
 
             // Legend datapoints
             let legendDataPoints = [
@@ -294,7 +329,7 @@ describe("Interactivity service", () => {
                 { selected: false, identity: selectableDataPoints[1].identity },
             ];
             let legendBehavior = new MockBehavior(legendDataPoints);
-            interactivityService.bind(legendDataPoints, legendBehavior, null, { isLegend: true });
+            interactivityService.bind({ dataPoints: legendDataPoints, behavior: legendBehavior, interactivityServiceOptions: { isLegend: true } });
 
             // Trigger selection on datapoints
             behavior.selectIndex(1);
@@ -324,7 +359,7 @@ describe("Interactivity service", () => {
                 { selected: false, identity: createSelectionId() },
             ];
             let legendBehavior = new MockBehavior(legendDataPoints);
-            interactivityService.bind(legendDataPoints, legendBehavior, null, { isLegend: true });
+            interactivityService.bind({ dataPoints: legendDataPoints, behavior: legendBehavior, interactivityServiceOptions: { isLegend: true } });
 
             // Select first legend item
             legendBehavior.selectIndex(0);
@@ -336,7 +371,7 @@ describe("Interactivity service", () => {
                 { selected: false, identity: createSelectionId() },
             ];
             legendBehavior = new MockBehavior(newLegendDataPoints);
-            interactivityService.bind(newLegendDataPoints, legendBehavior, null, { isLegend: true });
+            interactivityService.bind({ dataPoints: newLegendDataPoints, behavior: legendBehavior, interactivityServiceOptions: { isLegend: true } });
 
             // Select a new legend item
             legendBehavior.selectIndex(0);
@@ -359,8 +394,8 @@ describe("Interactivity service", () => {
                 { selected: false, identity: selectableDataPoints[1].identity },
             ];
             let labelBehavior = new MockBehavior(labelsDataPoints);
-            interactivityService.bind(selectableDataPoints, behavior, null);
-            interactivityService.bind(labelsDataPoints, labelBehavior, null, { isLabels: true });
+            interactivityService.bind({ dataPoints: selectableDataPoints, behavior: behavior });
+            interactivityService.bind({ dataPoints: labelsDataPoints, behavior: labelBehavior, interactivityServiceOptions: { isLabels: true } });
 
             labelBehavior.selectIndex(0);
             labelBehavior.verifySingleSelectedAt(0);