+++ CHANGELOG.md
@@ -1,5 +1,23 @@
+## 4.1.0
+* API 2.1.0 support
+
+## 4.x.x
+* Webpack support
+
+## 3.2.0
+* UPD: ability to select multiple data points via `handleSelection` method
+
+## 3.1.0
+* ADD: ability to support Power BI bookmarks
+
 ## 3.0.1
-* Moved to webpack 3, commonjs style.
+* Rename `applySelectionFromFitler` to `applySelectionFromFilter`
+
+## 3.0.0
+* New restoreFilter method of FilterManager class for converting AppliedFilter to IFilter (Basic or Advanced filter)
+
+## 2.0.2
+* New FilterManager class with restoreSelectionIds method for converting AppliedFilter to ISelectionId
 
 ## 2.0.1
 * UPD: syncSelectionState function declaration was added to ISelectionHandler interface